ktask/crudapi/requirements.txt
requests
mongoengine
python-decouple